Separate import logic into custom hooks
