+++ components/Transactions/Import/index.tsx
@@ -1,38 +1,19 @@
 import { Button, ListRow } from 'components/Styled';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
-import axios from 'axios';
 
 import styled from 'styled-components';
 import { useEffect, useState } from 'react';
-import { Transaction as PlaidTransaction } from 'plaid';
-import { useAlert } from 'components/Alert';
+import { useGetTransactionsFromBank } from 'lib/hooks/transaction/useGetTransactionsFromBank';
+import { useImportBankTransactions } from 'lib/hooks/transaction/useImportBankTransactions';
 const List = styled.div`
   display: flex;
   flex-direction: column;
 `;
 
-async function fetchTransactions() {
-  const response = await axios.get<PlaidTransaction[]>('/api/plaid/get_transactions');
-  return response.data;
-}
-
-async function postTransactions(transactions: PlaidTransaction[]) {
-  const response = await axios.post<PlaidTransaction[]>('/api/transactions', transactions);
-  return response.data;
-}
-
 export function Import() {
   const [fetchTrans, setFetchTrans] = useState(false);
-  const queryClient = useQueryClient();
-
-  const { addAlert } = useAlert();
   const [selectedTransactions, setSelectedTransactions] = useState([]);
-  const { data: transactions, isLoading } = useQuery('fetch-transactions-from-bank', fetchTransactions, {
-    enabled: fetchTrans,
-    staleTime: 10800000, // 3hr
-  });
-
-  const { mutate: uploadTransactions } = useMutation(postTransactions);
+  const { data: transactions } = useGetTransactionsFromBank({ enabled: fetchTrans });
+  const { mutate: uploadTransactions } = useImportBankTransactions();
 
   useEffect(() => {
     if (transactions) {
@@ -51,12 +32,7 @@ export function Import() {
     const selected = transactions.filter(
       transaction => selectedTransactions.indexOf(transaction.transaction_id) !== -1
     );
-    uploadTransactions(selected, {
-      onSuccess: () => {
-        addAlert({ message: 'Success!', type: 'success' });
-        queryClient.refetchQueries('fetch-transactions-from-bank');
-      },
-    });
+    uploadTransactions(selected);
   };
 
   if (!transactions) {
@@ -72,7 +48,7 @@ export function Import() {
   return (
     <div style={{ width: '100%' }}>
       <Button category="ACTION" onClick={handleUpload}>
-        Save Selected
+        Import Selected
       </Button>
       <Button category="PRIMARY" onClick={() => setSelectedTransactions([])}>
         Unselect All
